refactor(runtime): extract browser CDP/WebSocket into librefang-runtime-browser
Phase 1 of the #3710 runtime split. The new crate owns BrowserManager,
BrowserSession, CDP/WebSocket protocol types, and the embedded
content-extraction JS. Tool dispatchers (browser_navigate, click, type,
…) stay in runtime as `browser_tools` because they depend on
`crate::web_fetch::check_ssrf` and `crate::web_content::wrap_external_content`.

The `browser` feature on librefang-runtime now activates the standalone
crate (`pub use librefang_runtime_browser as browser`); disabling it
falls back to `browser_stub.rs`. Dispatcher callers updated to
`crate::browser_tools::tool_browser_*`.
