+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "crates/librefang-memory-wiki",
     "crates/librefang-runtime",
     "crates/librefang-runtime-audit",
+    "crates/librefang-runtime-browser",
     "crates/librefang-runtime-sandbox-docker",
     "crates/librefang-runtime-mcp",
     "crates/librefang-llm-drivers",
