+++ crates/librefang-runtime-browser/src/lib.rs
@@ -384,7 +384,7 @@ impl BrowserSession {
             let new_url = format!("{base}/json/new");
             let resp = tokio::time::timeout(
                 Duration::from_secs(CDP_CONNECT_TIMEOUT_SECS),
-                crate::http_client::new_client().get(&new_url).send(),
+                librefang_http::new_client().get(&new_url).send(),
             )
             .await
             .map_err(|_| format!("Timed out connecting to CDP endpoint: {cdp_endpoint}"))?
@@ -489,7 +489,7 @@ impl BrowserSession {
             if attempt > 0 {
                 tokio::time::sleep(Duration::from_millis(300)).await;
             }
-            let resp = match crate::http_client::new_client().get(list_url).send().await {
+            let resp = match librefang_http::new_client().get(list_url).send().await {
                 Ok(r) => r,
                 Err(_) => continue,
             };
@@ -971,10 +971,7 @@ impl BrowserManager {
                     let cdp_endpoint = self.config.cdp_endpoint.as_deref().unwrap_or("");
                     let base = cdp_endpoint.trim_end_matches('/');
                     let close_url = format!("{base}/json/close/{target_id}");
-                    let _ = crate::http_client::new_client()
-                        .get(&close_url)
-                        .send()
-                        .await;
+                    let _ = librefang_http::new_client().get(&close_url).send().await;
                     debug!(agent_id, target_id, "Closed remote CDP tab");
                 }
             }
@@ -1021,264 +1018,10 @@ impl BrowserManager {
     }
 }
 
-// ── Tool handler functions ─────────────────────────────────────────────────
-
-/// browser_navigate: Navigate to a URL. SSRF-checked before sending.
-pub async fn tool_browser_navigate(
-    input: &serde_json::Value,
-    mgr: &BrowserManager,
-    agent_id: &str,
-) -> Result<String, String> {
-    let url = input["url"].as_str().ok_or("Missing 'url' parameter")?;
-    // Browser navigation goes through CDP/WebSocket, not reqwest, so DNS-pinning
-    // the resolved address is not possible here. We still call check_ssrf to
-    // validate the URL scheme and reject IPs that resolve to internal/loopback
-    // ranges; the SsrfResolution result is intentionally discarded.
-    let _ = crate::web_fetch::check_ssrf(url, &[])?;
-
-    let resp = mgr
-        .send_command(
-            agent_id,
-            BrowserCommand::Navigate {
-                url: url.to_string(),
-            },
-        )
-        .await?;
-    if !resp.success {
-        return Err(resp.error.unwrap_or_else(|| "Navigate failed".to_string()));
-    }
-
-    let data = resp.data.unwrap_or_default();
-    let title = data["title"].as_str().unwrap_or("(no title)");
-    let page_url = data["url"].as_str().unwrap_or(url);
-    let content = data["content"].as_str().unwrap_or("");
-    let wrapped = crate::web_content::wrap_external_content(page_url, content);
-
-    Ok(format!(
-        "Navigated to: {page_url}\nTitle: {title}\n\n{wrapped}"
-    ))
-}
-
-/// browser_click: Click an element by CSS selector or visible text.
-pub async fn tool_browser_click(
-    input: &serde_json::Value,
-    mgr: &BrowserManager,
-    agent_id: &str,
-) -> Result<String, String> {
-    let selector = input["selector"]
-        .as_str()
-        .ok_or("Missing 'selector' parameter")?;
-
-    let resp = mgr
-        .send_command(
-            agent_id,
-            BrowserCommand::Click {
-                selector: selector.to_string(),
-            },
-        )
-        .await?;
-    if !resp.success {
-        return Err(resp.error.unwrap_or_else(|| "Click failed".to_string()));
-    }
-
-    let data = resp.data.unwrap_or_default();
-    let title = data["title"].as_str().unwrap_or("(no title)");
-    let url = data["url"].as_str().unwrap_or("");
-    Ok(format!("Clicked: {selector}\nPage: {title}\nURL: {url}"))
-}
-
-/// browser_type: Type text into an input field.
-pub async fn tool_browser_type(
-    input: &serde_json::Value,
-    mgr: &BrowserManager,
-    agent_id: &str,
-) -> Result<String, String> {
-    let selector = input["selector"]
-        .as_str()
-        .ok_or("Missing 'selector' parameter")?;
-    let text = input["text"].as_str().ok_or("Missing 'text' parameter")?;
-
-    let resp = mgr
-        .send_command(
-            agent_id,
-            BrowserCommand::Type {
-                selector: selector.to_string(),
-                text: text.to_string(),
-            },
-        )
-        .await?;
-    if !resp.success {
-        return Err(resp.error.unwrap_or_else(|| "Type failed".to_string()));
-    }
-    Ok(format!("Typed into {selector}: {text}"))
-}
-
-/// browser_screenshot: Take a screenshot of the current page.
-pub async fn tool_browser_screenshot(
-    _input: &serde_json::Value,
-    mgr: &BrowserManager,
-    agent_id: &str,
-    upload_dir: &std::path::Path,
-) -> Result<String, String> {
-    let resp = mgr
-        .send_command(agent_id, BrowserCommand::Screenshot)
-        .await?;
-    if !resp.success {
-        return Err(resp
-            .error
-            .unwrap_or_else(|| "Screenshot failed".to_string()));
-    }
-
-    let data = resp.data.unwrap_or_default();
-    let b64 = data["image_base64"].as_str().unwrap_or("");
-    let url = data["url"].as_str().unwrap_or("");
-
-    let mut image_urls: Vec<String> = Vec::new();
-    if !b64.is_empty() {
-        use base64::Engine;
-        let _ = std::fs::create_dir_all(upload_dir);
-        let file_id = uuid::Uuid::new_v4().to_string();
-        if let Ok(decoded) = base64::engine::general_purpose::STANDARD.decode(b64) {
-            let path = upload_dir.join(&file_id);
-            if std::fs::write(&path, &decoded).is_ok() {
-                image_urls.push(format!("/api/uploads/{file_id}"));
-            }
-        }
-    }
-
-    Ok(serde_json::json!({
-        "screenshot": true,
-        "url": url,
-        "image_urls": image_urls,
-    })
-    .to_string())
-}
-
-/// browser_read_page: Read current page content as markdown.
-pub async fn tool_browser_read_page(
-    _input: &serde_json::Value,
-    mgr: &BrowserManager,
-    agent_id: &str,
-) -> Result<String, String> {
-    let resp = mgr.send_command(agent_id, BrowserCommand::ReadPage).await?;
-    if !resp.success {
-        return Err(resp.error.unwrap_or_else(|| "ReadPage failed".to_string()));
-    }
-
-    let data = resp.data.unwrap_or_default();
-    let title = data["title"].as_str().unwrap_or("(no title)");
-    let url = data["url"].as_str().unwrap_or("");
-    let content = data["content"].as_str().unwrap_or("");
-    let wrapped = crate::web_content::wrap_external_content(url, content);
-
-    Ok(format!("Page: {title}\nURL: {url}\n\n{wrapped}"))
-}
-
-/// browser_close: Close the browser session.
-pub async fn tool_browser_close(
-    _input: &serde_json::Value,
-    mgr: &BrowserManager,
-    agent_id: &str,
-) -> Result<String, String> {
-    mgr.close_session(agent_id).await;
-    Ok("Browser session closed.".to_string())
-}
-
-/// browser_scroll: Scroll the page in a direction.
-pub async fn tool_browser_scroll(
-    input: &serde_json::Value,
-    mgr: &BrowserManager,
-    agent_id: &str,
-) -> Result<String, String> {
-    let direction = input["direction"].as_str().unwrap_or("down").to_string();
-    let amount = input["amount"].as_i64().unwrap_or(600) as i32;
-
-    let resp = mgr
-        .send_command(agent_id, BrowserCommand::Scroll { direction, amount })
-        .await?;
-    if !resp.success {
-        return Err(resp.error.unwrap_or_else(|| "Scroll failed".to_string()));
-    }
-    let data = resp.data.unwrap_or_default();
-    Ok(format!(
-        "Scrolled. Position: scrollX={}, scrollY={}",
-        data["scrollX"], data["scrollY"]
-    ))
-}
-
-/// browser_wait: Wait for a CSS selector to appear on the page.
-pub async fn tool_browser_wait(
-    input: &serde_json::Value,
-    mgr: &BrowserManager,
-    agent_id: &str,
-) -> Result<String, String> {
-    let selector = input["selector"]
-        .as_str()
-        .ok_or("Missing 'selector' parameter")?;
-    let timeout_ms = input["timeout_ms"].as_u64().unwrap_or(5000);
-
-    let resp = mgr
-        .send_command(
-            agent_id,
-            BrowserCommand::Wait {
-                selector: selector.to_string(),
-                timeout_ms,
-            },
-        )
-        .await?;
-    if !resp.success {
-        return Err(resp.error.unwrap_or_else(|| "Wait timed out".to_string()));
-    }
-    Ok(format!("Element found: {selector}"))
-}
-
-/// browser_run_js: Run JavaScript on the current page.
-pub async fn tool_browser_run_js(
-    input: &serde_json::Value,
-    mgr: &BrowserManager,
-    agent_id: &str,
-) -> Result<String, String> {
-    let expression = input["expression"]
-        .as_str()
-        .ok_or("Missing 'expression' parameter")?;
-
-    let resp = mgr
-        .send_command(
-            agent_id,
-            BrowserCommand::RunJs {
-                expression: expression.to_string(),
-            },
-        )
-        .await?;
-    if !resp.success {
-        return Err(resp
-            .error
-            .unwrap_or_else(|| "JS execution failed".to_string()));
-    }
-    let data = resp.data.unwrap_or_default();
-    Ok(serde_json::to_string_pretty(&data["result"]).unwrap_or_else(|_| "null".to_string()))
-}
-
-/// browser_back: Go back in browser history.
-pub async fn tool_browser_back(
-    _input: &serde_json::Value,
-    mgr: &BrowserManager,
-    agent_id: &str,
-) -> Result<String, String> {
-    let resp = mgr.send_command(agent_id, BrowserCommand::Back).await?;
-    if !resp.success {
-        return Err(resp.error.unwrap_or_else(|| "Back failed".to_string()));
-    }
-    let data = resp.data.unwrap_or_default();
-    let title = data["title"].as_str().unwrap_or("(no title)");
-    let url = data["url"].as_str().unwrap_or("");
-    Ok(format!("Went back.\nPage: {title}\nURL: {url}"))
-}
-
 // ── Embedded JavaScript ────────────────────────────────────────────────────
 
 /// JavaScript to extract readable page content as markdown.
-const EXTRACT_CONTENT_JS: &str = r#"(() => {
+pub(crate) const EXTRACT_CONTENT_JS: &str = r#"(() => {
     const title = document.title || '';
     const url = location.href || '';
     const body = document.body;
@@ -1481,3 +1224,5 @@ mod tests {
         assert_eq!(err.error.unwrap(), "bad");
     }
 }
+
+// ── Tool handler functions ─────────────────────────────────────────────────
