+++ crates/librefang-runtime/Cargo.toml
@@ -11,6 +11,7 @@ librefang-types = { path = "../librefang-types" }
 librefang-http = { path = "../librefang-http" }
 librefang-kernel-handle = { path = "../librefang-kernel-handle" }
 librefang-runtime-audit = { path = "../librefang-runtime-audit" }
+librefang-runtime-browser = { path = "../librefang-runtime-browser", optional = true }
 librefang-runtime-sandbox-docker = { path = "../librefang-runtime-sandbox-docker", optional = true }
 librefang-runtime-mcp = { path = "../librefang-runtime-mcp" }
 librefang-llm-drivers = { path = "../librefang-llm-drivers" }
@@ -96,7 +97,7 @@ daytona-backend = []
 # modules and consumer code paths so build-time configurations can omit
 # whole subsystems for size/security.
 media = []
-browser = []
+browser = ["dep:librefang-runtime-browser"]
 docker-sandbox = ["dep:librefang-runtime-sandbox-docker"]
